+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Spacing units calculations internally. Mobile layouts improved.
+
+### Added
+- `compact` prop for `Box`, `Panel`, and `Stack`.
 
 ## [v0.7.0]
 ### Changed